Renaming and microfixes
