+++ src/nncf/experimental/quantization/quantizer.py
@@ -11,50 +11,12 @@
 
 from abc import ABC
 from abc import abstractmethod
-from enum import Enum
-from typing import Any, Optional, TypeVar
+from typing import TypeVar
 
 from nncf.common.graph.graph import NNCFGraph
 from nncf.common.quantization.quantizer_setup import SingleConfigQuantizerSetup
-from nncf.common.quantization.structs import QuantizationScheme
-from nncf.common.quantization.structs import QuantizerConfig
-from nncf.config.schemata.defaults import QUANTIZATION_BITS
-from nncf.config.schemata.defaults import QUANTIZATION_NARROW_RANGE
-from nncf.config.schemata.defaults import QUANTIZATION_PER_CHANNEL
 
 TModel = TypeVar("TModel")
-
-
-class IntDtype(Enum):
-    """
-    Enum of possible integer types.
-    """
-
-    INT8 = "INT8"
-    UINT8 = "UINT8"
-
-
-class FXQuantizerConfig(QuantizerConfig):
-    def __init__(
-        self,
-        num_bits: int = QUANTIZATION_BITS,
-        mode: QuantizationScheme = QuantizationScheme.SYMMETRIC,
-        signedness_to_force: Optional[bool] = None,
-        per_channel: bool = QUANTIZATION_PER_CHANNEL,
-        narrow_range: bool = QUANTIZATION_NARROW_RANGE,
-        dest_dtype: IntDtype = IntDtype.INT8,
-    ):
-        super().__init__(num_bits, mode, signedness_to_force, per_channel, narrow_range)
-        self.dest_dtype = dest_dtype
-
-    def __str__(self) -> str:
-        retval = super().__str__()
-        return retval + " DestDtype: {self._dest_dtype}"
-
-    def get_state(self) -> dict[str, Any]:
-        state = super().get_state()
-        state["dest_dtype"] = self.dest_dtype
-        return state
 
 
 class Quantizer(ABC):
